+++ CHANGELOG.md
@@ -11,6 +11,7 @@ For a full diff see [`2.22.2...main`][2.22.2...main].
 ### Changed
 
 - Added table headers to slow test report output ([#768]), by [@localheinz]
+- Rendered slow test report with a duration format legend ([#773]), by [@localheinz]
 
 ## [`2.22.2`][2.22.2]
 
@@ -482,6 +483,7 @@ For a full diff see [`7afa59c...1.0.0`][7afa59c...1.0.0].
 [#762]: https://github.com/ergebnis/phpunit-slow-test-detector/pull/762
 [#765]: https://github.com/ergebnis/phpunit-slow-test-detector/pull/765
 [#768]: https://github.com/ergebnis/phpunit-slow-test-detector/pull/768
+[#773]: https://github.com/ergebnis/phpunit-slow-test-detector/pull/773
 
 [@courtney-miles]: https://github.com/courtney-miles
 [@dantleech]: https://github.com/dantleech